Add simple external API for projects
In order to demonstrate RSpec request specs, I'd like a simple,
standalone API controller that leverages the token-based auth created
previously. Add an API for accessing a subset of projects-related
features (index, show, and create) using a URL structure like:

  http://localhost:3000/api/projects.json?user_email=me@example.com&user_token=mytoken

I know this isn't a super-robust example, but as I say in the book, this
isn't a book about designing robust APIs, it's a book about testing. I
may create a more complex example in the future, but not right now.