Add request spec examples
The sample app provides a simple authenticated API for viewing and
creating projects. Add a couple of example tests to cover this, along
with an example of how to use request specs to test regular Rails
controllers. Include a helper module (cribbed from the Devise wiki) to
allow Devise logins in the second example.