Remove unnecessary test data from tasks controller specs
The examples in the "describe #create" block of the tasks controller
spec pass an extra "id" param, which is not needed for the create
action. This comes to light when switching the spec to create test data
with "let" instead of an instance variable, for chapter 8.

Remove the extra data from the params passed to the create action in the
examples.

Referenced in
https://github.com/everydayrails/everydayrails-rspec-2017/issues/4