Reformat failure messages in custom matcher
After generating chapter 8's chapter in Leanpub preview mode, I saw that
the failure messages in the custom matcher didn't wrap well. After
fiddling with ways to keep using %Q() to define the strings, I switched
to concatenating the string with + and breaking across multiple lines.
It's not quite as elegant, but wraps correctly.