Re-record VCR cassette following update to Geocoder
Geocoder now uses the IPInfo.io service as its default for IP-based
geocoding, since the previous default service is no longer available.
This breaks test examples in chapter 10 of Everyday Rails Testing with
RSpec, which uses a geocoding example to record a VCR cassette of
network activity.

Re-record the cassette to use IPInfo.io. It looks like this service's
response is slightly different than the previous service, so update the
expected response to reflect that.

IPInfo.io doesn't require an API key for less than 1,000 requests a day,
and is limited to non-HTTPS requests only. These limitations should be
adequate for the sake of demonstration in the book, but should not be
used as-is in a production environment. See IPInfo.io's pricing page to
acquire an API key for the service.

In addition, you'll want to ensure that secrets like API keys are
filtered from cassettes when using VCR to record any service requiring
an API key, username/password, or other secret. See
https://relishapp.com/vcr/vcr/v/3-0-3/docs/configuration/filter-sensitive-data
for details on filtering sensitive data in the version of VCR used in
the current version of  Everyday Rails Testing with RSpec.

https://github.com/everydayrails/everydayrails-rspec-2017/issues/84